requirements.txt
discord.py==2.7.0
google-genai
groq
mistralai
cohere
python-dotenv==1.2.1
requests==2.32.5
feedparser
aiohttp
asyncio
yt-dlp
ffmpeg-python
davey
PyNaCl
